Perhaps temporary fix:  getResourceZipAsFile()

+++ connectionUtils/src/main/java/com/ge/research/semtk/utility/Utility.java
@@ -76,6 +76,7 @@ import javax.xml.parsers.DocumentBuilder;
 import javax.xml.parsers.DocumentBuilderFactory;
 
 import org.apache.commons.collections.map.CaseInsensitiveMap;
+import org.apache.commons.compress.utils.FileNameUtils;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVPrinter;
 import org.apache.commons.io.FileUtils;
@@ -865,22 +866,54 @@ public abstract class Utility {
 		return ret;
 	}
 	
-	@Deprecated // Use getResourceAsStream
 	public static File getResourceAsFile(Object obj, String fileName) throws Exception {
 		return getResourceAsFile(obj.getClass(), fileName);
 	}
-	@Deprecated // Use getResourceAsStream    see end of https://stackoverflow.com/questions/18055189/why-is-my-uri-not-hierarchical
 	public static File getResourceAsFile(Class c, String fileName) throws Exception {
 
+		String ext = FileNameUtils.getExtension(fileName).toLowerCase();
+		if (ext.equals("zip")) return getResourceAsZipFile(c, fileName);
+		
 		File ret = null;
 		
 		// fat jars and other deployments seem to choke on returning a File
 		// so copy the data into a temp file
+		// Resources should be retrieved as Streams, not files.
 		byte data[] = null;
 		data = getResourceAsBytes(c, fileName);
-		File tempFile = File.createTempFile("resource", "fileName");
+		File tempFile = File.createTempFile("resource", "." + FileNameUtils.getExtension(fileName));
 		tempFile.deleteOnExit();
 		FileUtils.writeByteArrayToFile(tempFile, data);
+		
+		return tempFile;
+	}
+	
+	/**
+	 * For some reason copying bytes in getResourceAsFile() does not work for zip files.
+	 * @param obj
+	 * @param fileName
+	 * @return
+	 * @throws Exception
+	 */
+	public static File getResourceZipAsFile(Object obj, String fileName) throws Exception {
+		return getResourceAsFile(obj.getClass(), fileName);
+	}
+	public static File getResourceAsZipFile(Class c, String fileName) throws Exception {
+		
+		InputStream in = c.getResourceAsStream(fixResourceName(fileName));
+		if (in == null) {
+			throw new Exception("Could find resource file: " + fileName);
+		}
+		ZipInputStream zin = new ZipInputStream(in);
+		
+		File tempFile = File.createTempFile("resource", ".zip");
+		ZipOutputStream zout = new ZipOutputStream(new FileOutputStream(tempFile));
+		ZipEntry e = null;
+		while ((e = zin.getNextEntry()) != null) {
+			zout.putNextEntry(new ZipEntry(e));
+		}
+		zin.close();
+		zout.close();
 		
 		return tempFile;
 	}
@@ -890,7 +923,7 @@ public abstract class Utility {
 	}
 	public static byte [] getResourceAsBytes(Class c, String fileName) throws Exception  {
 		byte [] ret = null;
-
+		
 		InputStream in = c.getResourceAsStream(fixResourceName(fileName));
 		if (in == null) {
 			throw new Exception("Could find resource file: " + fileName);
